pricing_agent/requirements.txt
flask==3.1.0
pillow==10.3.0
requests==2.32.3
gunicorn==21.2.0
openai==1.59.7
beautifulsoup4==4.12.3
python-dotenv==1.0.1
flask-wtf==1.2.1
flask-sqlalchemy==3.1.1